model1/requirements.txt
streamlit
pandas
numpy
plotly
altair
matplotlib
python-dotenv
gemini
